requirements.txt
Django
psycopg2-binary
Pillow
djangorestframework
python-decouple
django-phonenumber-field
phonenumbers
twilio
djangorestframework-simplejwt
drf-yasg
django-filter